+++ doc/examples/gestures/lib/main_mouse_movement.dart
@@ -1,13 +1,16 @@
-import 'package:flutter/material.dart';
 import 'package:flame/game.dart';
+import 'package:flutter/material.dart';
 import 'package:flame/gestures.dart';
 import 'package:flame/palette.dart';
 import 'package:flame/extensions/vector2.dart';
 import 'package:flame/extensions/offset.dart';
 
 void main() {
-  final game = MyGame();
-  runApp(game.widget);
+  runApp(
+    GameWidget(
+      game: MyGame(),
+    ),
+  );
 }
 
 class MyGame extends Game with MouseMovementDetector {
