+++ doc/examples/gestures/lib/main_tapables.dart
@@ -6,15 +6,15 @@ import 'package:flame/components/position_component.dart';
 import 'package:flame/components/mixins/tapable.dart';
 
 void main() {
-  final game = MyGame();
-
-  final widget = Container(
-    padding: const EdgeInsets.all(50),
-    color: const Color(0xFFA9A9A9),
-    child: game.widget,
+  runApp(
+    Container(
+      padding: const EdgeInsets.all(50),
+      color: const Color(0xFFA9A9A9),
+      child: GameWidget(
+        game: MyGame(),
+      ),
+    ),
   );
-
-  runApp(widget);
 }
 
 class TapableSquare extends PositionComponent with Tapable {
