+++ doc/examples/go_desktop/lib/main.dart
@@ -1,8 +1,13 @@
+import 'package:flame/game.dart';
 import 'package:flutter/material.dart';
 
 import './game.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
-  runApp(MyGame().widget);
+  runApp(
+    GameWidget(
+      game: MyGame(),
+    ),
+  );
 }
