+++ doc/examples/go_desktop/lib/main_desktop.dart
@@ -1,3 +1,4 @@
+import 'package:flame/game.dart';
 import 'package:flutter/foundation.dart'
     show debugDefaultTargetPlatformOverride;
 import 'package:flutter/material.dart';
@@ -7,5 +8,9 @@ import './game.dart';
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
   debugDefaultTargetPlatformOverride = TargetPlatform.fuchsia;
-  runApp(MyGame().widget);
+  runApp(
+    GameWidget(
+      game: MyGame(),
+    ),
+  );
 }
