+++ doc/examples/keyboard/lib/main.dart
@@ -5,7 +5,11 @@ import 'package:flutter/services.dart';
 
 import 'dart:ui';
 
-void main() => runApp(MyGame().widget);
+void main() => runApp(
+      GameWidget(
+        game: MyGame(),
+      ),
+    );
 
 class MyGame extends Game with KeyboardEvents {
   static final Paint _white = Paint()..color = const Color(0xFFFFFFFF);
