+++ doc/examples/nine_tile_box/lib/main.dart
@@ -11,7 +11,11 @@ void main() async {
   final size = await Flame.util.initialDimensions();
 
   final game = MyGame(size);
-  runApp(game.widget);
+  runApp(
+    GameWidget(
+      game: game,
+    ),
+  );
 }
 
 class MyGame extends Game {
