+++ doc/examples/sprites/lib/main_base64.dart
@@ -1,13 +1,16 @@
 import 'package:flame/extensions/vector2.dart';
-import 'package:flutter/material.dart';
-
-import 'package:flame/sprite.dart';
 import 'package:flame/game.dart';
+import 'package:flutter/material.dart';
+import 'package:flame/sprite.dart';
 
 import 'dart:ui';
 
 void main() {
-  runApp(MyGame().widget);
+  runApp(
+    GameWidget(
+      game: MyGame(),
+    ),
+  );
 }
 
 class MyGame extends Game {
