+++ lib/game.dart
@@ -1,3 +1,4 @@
 // Keeping compatible with earlier versions of Flame
 export './game/base_game.dart';
 export './game/game.dart';
+export './game/game_widget.dart';
